todo_item/serializers.py
from rest_framework import serializers
from rest_framework.serializers import ModelSerializer
from .models import TodoItem

class TodoItemSerializer(ModelSerializer):
	class Meta:
		model = TodoItem
		fields = '__all__'