requirements.txt
djangorestframework
Django
bcrypt
gunicorn==19.6.0
whitenoise
